Extract invitations package
