+++ src/main/java/com/code4ro/legalconsultation/invitation/model/persistence/Invitation.java
@@ -1,6 +1,7 @@
-package com.code4ro.legalconsultation.model.persistence;
+package com.code4ro.legalconsultation.invitation.model.persistence;
 
 
+import com.code4ro.legalconsultation.model.persistence.BaseEntity;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 import lombok.Getter;
 import lombok.Setter;
