+++ src/main/java/com/code4ro/legalconsultation/invitation/model/persistence/InvitationStatus.java
@@ -1,0 +1,5 @@
+package com.code4ro.legalconsultation.invitation.model.persistence;
+
+public enum InvitationStatus {
+    PENDING, USED, EXPIRED
+}
