+++ src/main/java/com/code4ro/legalconsultation/invitation/service/InvitationService.java
@@ -1,7 +1,7 @@
-package com.code4ro.legalconsultation.service.api;
+package com.code4ro.legalconsultation.invitation.service;
 
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
-import com.code4ro.legalconsultation.model.persistence.Invitation;
+import com.code4ro.legalconsultation.invitation.model.persistence.Invitation;
 import com.code4ro.legalconsultation.user.model.persistence.User;
 
 public interface InvitationService {
