--- src/main/java/com/code4ro/legalconsultation/model/persistence/InvitationStatus.java
@@ -1,5 +1,0 @@
-package com.code4ro.legalconsultation.model.persistence;
-
-public enum InvitationStatus {
-    PENDING, USED, EXPIRED
-}
