+++ src/test/java/com/code4ro/legalconsultation/comment/factory/CommentFactory.java
@@ -1,9 +1,10 @@
-package com.code4ro.legalconsultation.factory;
+package com.code4ro.legalconsultation.comment.factory;
 
 import com.code4ro.legalconsultation.comment.model.dto.CommentDetailDto;
 import com.code4ro.legalconsultation.comment.model.dto.CommentDto;
 import com.code4ro.legalconsultation.comment.model.persistence.Comment;
 import com.code4ro.legalconsultation.comment.service.CommentService;
+import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
 import org.apache.commons.lang3.RandomStringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
