+++ src/test/java/com/code4ro/legalconsultation/common/controller/AbstractControllerIntegrationTest.java
@@ -2,10 +2,10 @@ package com.code4ro.legalconsultation.common.controller;
 
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
-import com.code4ro.legalconsultation.model.persistence.Invitation;
-import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
+import com.code4ro.legalconsultation.invitation.model.persistence.Invitation;
+import com.code4ro.legalconsultation.invitation.model.persistence.InvitationStatus;
 import com.code4ro.legalconsultation.user.model.persistence.User;
-import com.code4ro.legalconsultation.repository.InvitationRepository;
+import com.code4ro.legalconsultation.invitation.repository.InvitationRepository;
 import com.code4ro.legalconsultation.user.repository.UserRepository;
 import com.code4ro.legalconsultation.service.impl.ApplicationUserService;
 import com.fasterxml.jackson.databind.ObjectMapper;
