+++ src/test/java/com/code4ro/legalconsultation/invitation/service/InvitationServiceTest.java
@@ -1,13 +1,13 @@
-package com.code4ro.legalconsultation.service;
+package com.code4ro.legalconsultation.invitation.service;
 
 import com.code4ro.legalconsultation.common.exceptions.LegalValidationException;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.SignUpRequest;
-import com.code4ro.legalconsultation.model.persistence.Invitation;
-import com.code4ro.legalconsultation.model.persistence.InvitationStatus;
+import com.code4ro.legalconsultation.invitation.model.persistence.Invitation;
+import com.code4ro.legalconsultation.invitation.model.persistence.InvitationStatus;
 import com.code4ro.legalconsultation.user.model.persistence.User;
-import com.code4ro.legalconsultation.repository.InvitationRepository;
-import com.code4ro.legalconsultation.service.impl.InvitationServiceImpl;
+import com.code4ro.legalconsultation.invitation.repository.InvitationRepository;
+import com.code4ro.legalconsultation.invitation.service.impl.InvitationServiceImpl;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.mockito.ArgumentCaptor;
