+++ src/test/java/com/code4ro/legalconsultation/service/impl/VoteServiceImplSaveOrUpdateTest.java
@@ -3,7 +3,7 @@ package com.code4ro.legalconsultation.service.impl;
 import com.code4ro.legalconsultation.comment.model.persistence.Comment;
 import com.code4ro.legalconsultation.comment.service.CommentService;
 import com.code4ro.legalconsultation.config.security.CurrentUserService;
-import com.code4ro.legalconsultation.factory.CommentFactory;
+import com.code4ro.legalconsultation.comment.factory.CommentFactory;
 import com.code4ro.legalconsultation.factory.RandomObjectFiller;
 import com.code4ro.legalconsultation.model.dto.VoteDto;
 import com.code4ro.legalconsultation.model.persistence.ApplicationUser;
